+++ lib/processor/sql.go
@@ -51,7 +51,7 @@ SQL is a processor that runs a query against a target database for each message
 batch and, for queries that return rows, replaces the batch with the result.
 
 In order to execute an SQL query for each message of the batch use this
-processor within a ` + "[`process_batch`](#process_batch)" + ` processor.
+processor within a ` + "[`for_each`](#for_each)" + ` processor.
 
 If a query contains arguments they can be set as an array of strings supporting
 [interpolation functions](../config_interpolation.md#functions) in the